+++ tests/runner.test.ts
@@ -3,7 +3,7 @@ import { ChildProcess } from 'child_process';
 import { assert } from 'chai';
 
 import config from './config/config';
-import * as yerbamate from '..';
+import * as yerbamate from '../main';
 
 const { run, stop } = yerbamate;
 
@@ -17,7 +17,7 @@ describe("Runner", () => {
     it("Execute script", (done) => {
         run("node " + path.join(testDir, testScript), (code, outs, errs) => {
             checkDefaultOutput(code, outs, errs);
-            assert.strictEqual(outs[1], "2");
+            assert.strictEqual(outs.split("\n")[1], "2");
             done();
         });
     });
@@ -25,7 +25,7 @@ describe("Runner", () => {
     it("Different directory", (done) => {
         run("node " + testScript, testDir, (code, outs, errs) => {
             checkDefaultOutput(code, outs, errs);
-            assert.strictEqual(outs[1], "2");
+            assert.strictEqual(outs.split("\n")[1], "2");
             done();
         });
     });
@@ -35,7 +35,7 @@ describe("Runner", () => {
             args: "myargument1 myargument2"
         }, function(code, outs, errs) {
             checkDefaultOutput(code, outs, errs);
-            assert.strictEqual(outs[1], "4");
+            assert.strictEqual(outs.split("\n")[1], "4");
             done();
         });
     });
@@ -47,11 +47,8 @@ describe("Runner", () => {
             assert.ok(outs);
             assert.ok(errs);
             assert.strictEqual(code, 1);
-            assert.lengthOf(outs, 2);
-            assert.isAtLeast(errs.length, 2);
-            assert.strictEqual(outs[0], "Example js running");
-            assert.strictEqual(errs[0], "Warning example");
-            assert.strictEqual(outs[1], "3");
+            assert.strictEqual(outs, "Example js running\n3\n");
+            assert.strictEqual(errs.split("\n")[0], "Warning example");
             done();
         });
     });
@@ -59,14 +56,13 @@ describe("Runner", () => {
     it("Invalid command", (done) => {
         run(testScript, testDir, (code, outs, errs) => {
             assert.notEqual(code, 0);
-            assert.lengthOf(outs, 0);
-            assert.lengthOf(errs, 1);
+            assert.strictEqual(outs, "");
+            assert.ok(errs);
             done();
         });
-
     });
 
-    it("Stout and stderr hooks", (done) => {
+    it("Stdout and stderr hooks", (done) => {
         let outTest = "";
         let errTest = "";
         const onOut = (data: any) => {
@@ -82,59 +78,60 @@ describe("Runner", () => {
             stdout: onOut,
             stderr: onErr
         }, function(code, outs, errs) {
-            const outTestArr = outTest.split('\n').filter(Boolean);
-            const errTestArr = errTest.split('\n').filter(Boolean);
             checkDefaultOutput(code, outs, errs);
-            assert.strictEqual(outs[1], "2");
-            checkDefaultOutput(code, outTestArr, errTestArr);
+            assert.strictEqual(outs.split("\n")[1], "2");
+            checkDefaultOutput(code, outTest, errTest);
             done();
         });
     });
 
-    it("Stop process", (done) => {
-        const proc = run("node", (code) => {
-            assert.notEqual(code, 0);
-            done();
-        }) as ChildProcess;
-        assert.ok(proc);
-        stop(proc);
-    });
-
-    it("Stop process callback", (done) => {
-        const proc = run("node", () => { }) as ChildProcess;
-        assert.ok(proc);
-        stop(proc, function(err: any) {
-            assert.notOk(err);
-            done();
+    describe("stop", () => {
+        it("Stop process", (done) => {
+            const proc = run("node", (code) => {
+                assert.notEqual(code, 0);
+                done();
+            }) as ChildProcess;
+            assert.ok(proc);
+            stop(proc);
         });
-    });
+
+        it("Stop process callback", (done) => {
+            const proc = run("node", () => { }) as ChildProcess;
+            assert.ok(proc);
+            stop(proc, function(err: any) {
+                assert.notOk(err);
+                done();
+            });
+        });
+
+        it("Throws if stop is called without pid", () => {
+            assert.throws(() => {
+                stop(undefined as any)
+            })
+            assert.throws(() => {
+                stop({ pid: undefined } as any)
+            })
+            assert.throws(() => {
+                stop({ pid: null } as any)
+            })
+        })
+    })
 
     it("Array arguments", (done) => {
         run("node " + testScript, testDir, {
             args: ["myargument1", "myargument2"]
         }, function(code, outs, errs) {
             checkDefaultOutput(code, outs, errs);
-            assert.strictEqual(outs[1], "4");
-            done();
-        });
-    });
-
-    it("Execute script with undefined options and arguments", (done) => {
-        run("node " + path.join(testDir, testScript), undefined, undefined, (code, outs, errs) => {
-            checkDefaultOutput(code, outs, errs);
-            assert.strictEqual(outs[1], "2");
+            assert.strictEqual(outs.split("\n")[1], "4");
             done();
         });
     });
 
     it("Execute envs defined in command", (done) => {
         run("testenv=dontpanic node print_env.js", testDir, (code, outs, errs) => {
-            assert.ok(outs);
-            assert.ok(errs);
             assert.strictEqual(code, 0);
-            assert.lengthOf(outs, 1);
-            assert.lengthOf(errs, 0);
-            assert.strictEqual(outs[0], "dontpanic");
+            assert.strictEqual(errs, "");
+            assert.strictEqual(outs, "dontpanic\n");
             done();
         });
     });
@@ -145,12 +142,9 @@ describe("Runner", () => {
                 testenv: "dontpanic"
             }
         }, (code, outs, errs) => {
-            assert.ok(outs);
-            assert.ok(errs);
             assert.strictEqual(code, 0);
-            assert.lengthOf(outs, 1);
             assert.lengthOf(errs, 0);
-            assert.strictEqual(outs[0], "dontpanic");
+            assert.strictEqual(outs, "dontpanic\n");
             done();
         });
     });
@@ -160,7 +154,7 @@ describe("Runner", () => {
             args: ["myargument1", "myargument2"]
         }, (code, outs, errs) => {
             checkDefaultOutput(code, outs, errs);
-            assert.strictEqual(outs[1], "4");
+            assert.strictEqual(outs.split("\n")[1], "4");
             done();
         });
     });
@@ -169,12 +163,15 @@ describe("Runner", () => {
         run("node " + path.join(testDir, testScript), {
             maxOutputSize: 10
         }, (_code, outs, errs) => {
-            assert.lengthOf(outs, 2);
-            assert.strictEqual(outs[0], "running");
-            assert.strictEqual(outs[1], "2");
-            assert.lengthOf(errs, 1);
-            assert.strictEqual(errs[0], "g example");
+            assert.strictEqual(outs, "running\n2\n");
+            assert.strictEqual(errs, "g example\n");
             done();
         });
     });
+
+    it("Throws if callback is not provided", ()=>{
+        assert.throws(()=>{
+            run("node "+ path.join(testDir, testScript), undefined as any);
+        })
+    })
 });