+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 4.0.0 / ####-##-##
 ==================
 
-* Using typescript
+* Stderr and stdout now return as a string
+* Typescript support
+* Process environment variables are always added
+* Environment variables set through settings will override process env variables
 * Exports changed to support TypeScript